[Mod] Fix typo in exception handler. [Fix] Truncating cast from HMENU to UINT - probably not a problem since Window handles are portable between 32-bit and 64-bit applications, so must fit into 32-bit anyway. [Ref] Silence more warnings.
git-svn-id: https://source.openmpt.org/svn/openmpt/trunk/OpenMPT@6386 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ mptrack/Mptrack.cpp
@@ -434,7 +434,7 @@ BOOL CTrackApp::ImportMidiConfig(SettingsContainer &file, bool forgetSettings)
 		GetCurrentDirectoryW(CountOf(curDir), curDir);
 		UltraSndPath = mpt::PathString::FromNative(curDir);
 	}
-	for (UINT iMidi=0; iMidi<256; iMidi++)
+	for (uint32 iMidi=0; iMidi<256; iMidi++)
 	{
 		mpt::PathString filename;
 		char section[32];
@@ -487,7 +487,7 @@ BOOL CTrackApp::ExportMidiConfig(const mpt::PathString &filename)
 BOOL CTrackApp::ExportMidiConfig(SettingsContainer &file)
 //-------------------------------------------------------
 {
-	for(size_t iMidi = 0; iMidi < 256; iMidi++) if (!midiLibrary.MidiMap[iMidi].empty())
+	for(uint32 iMidi = 0; iMidi < 256; iMidi++) if (!midiLibrary.MidiMap[iMidi].empty())
 	{
 		mpt::PathString szFileName = midiLibrary.MidiMap[iMidi];
